Chapter 02: change type from double to unsigned int as instructed by the text

+++ Examples/NoModules/Chapter 02/Ex2_03A.cpp
@@ -1,5 +1,6 @@
 // Expressions with mixed variables types
-// (The difference with the original example is the type of fish_length)
+// (The difference with the original example 
+//  is the type of fish_count and inches_per_foot)
 #include <iostream>
 #include <numbers>   // For the pi constant
 #include <cmath>     // For the square root function
@@ -7,11 +8,11 @@
 int main()
 {
   // 2 square feet pond surface for every 6 inches of fish
-  const double fish_factor { 2.0/0.5 };  // Area per unit length of fish
-  const double inches_per_foot { 12.0 };
+  const double fish_factor { 2.0/0.5 };       // Area per unit length of fish
+  const unsigned int inches_per_foot { 12 };  // <-- Used to be of type double
 
-  unsigned int fish_count {};      // Number of fish
-  double fish_length {};           // Average length of fish
+  unsigned int fish_count {};  // Number of fish (used to be of type double as well)
+  double fish_length {};       // Average length of fish
 
   std::cout << "Enter the number of fish you want to keep: ";
   std::cin >> fish_count;
